+++ root_orchestrator/resource-abstractor/api/v1/hooks_blueprint.py
@@ -1,5 +1,3 @@
-from enum import Enum
-
 from db import hooks_db
 from flask import request
 from flask.views import MethodView
@@ -9,23 +7,14 @@ from marshmallow import Schema, fields, validate
 hooksblp = Blueprint("Hooks Api", "hooks_api", url_prefix="/api/v1/hooks")
 
 
-class HookEventsEnum(Enum):
-    AFTER_CREATE = "afterCreate"
-    BEFORE_CREATE = "beforeCreate"
-
-    AFTER_UPDATE = "afterUpdate"
-    BEFORE_UPDATE = "beforeUpdate"
-
-    AFTER_DELETE = "afterDelete"
-
-
 class APIObjectHookSchema(Schema):
     _id = fields.String()
     hook_name = fields.String(required=True)
     webhook_url = fields.String(required=True)
     entity = fields.String(required=True)
     events = fields.List(
-        fields.Str(validate=validate.OneOf([e.value for e in HookEventsEnum])), required=True
+        fields.Str(validate=validate.OneOf([e.value for e in hooks_db.HookEventsEnum])),
+        required=True,
     )
