+++ root_orchestrator/resource-abstractor/managers/hook_manager.py
@@ -4,14 +4,16 @@ from requests import exceptions, post
 
 def before_create_decorator(func):
     def wrapper(data, type):
-        data = beforeObjectCreation(data, type)
+        data = hook_before_creation(data, type)
         return func(data, type)
 
     return wrapper
 
 
-def beforeObjectCreation(data, type):
-    hooks = hooks_db.mongo_get_hooks({"entity": type, "events": "beforeCreate"})
+def hook_before_creation(data, type):
+    hooks = hooks_db.mongo_get_hooks(
+        {"entity": type, "events": {"$in": [hooks_db.HookEventsEnum.BEFORE_CREATE]}}
+    )
     for hook in hooks:
         try:
             response = post(hook["webhook_url"], json=data)
